PM-17 added rest CRUD for pets and refactored db instance
