+++ package.json
@@ -24,6 +24,7 @@
   },
   "devDependencies": {
     "chai": "^4.3.6",
+    "mocha": "^9.2.0",
     "nodemon": "^2.0.15",
     "supertest": "^6.0.1"
   }
