+++ src/config/db.config.js
@@ -4,5 +4,10 @@ module.exports = {
   PORT: '3306',
   USER: "[YOUR_USERNAME]",
   PASSWORD: "[YOUR_PASSWORD]",
-  DB: "purrfect-match_dev"
+  DB: "purrfect-match_dev",
+  DEV_HOST: "localhost",
+  DEV_PORT: '3306',
+  DEV_USER: "esther",
+  DEV_PASSWORD: "P@ssw0rd",
+  DEV_DB: "purrfect-match_dev"
 };
