+++ src/controllers/sheltersController.js
@@ -23,7 +23,7 @@ const postShelters = async (req, res, next) => {
   }
 
   if (success) {
-    await createShelters(req.body)
+    await createShelters(req.app.get('db'), req.body)
       .then((dbResponse) => {
         res.status(201).type('json').send(dbResponse);
       })
@@ -39,7 +39,7 @@ const postShelters = async (req, res, next) => {
 
 const getShelter = async (req, res, next) => {
   log.debug("Calling getShelter...");
-  await getShelterByID(req.params.shelterID)
+  await getShelterByID(req.app.get('db'), req.params.shelterID)
     .then((dbResponse) => {
       if (dbResponse.length == 0) res.sendStatus(404);
       else res.send(dbResponse);
@@ -65,7 +65,7 @@ const updateShelter = async (req, res, next) => {
   }
 
   if (success) {
-    await updateShelterByID(req.params.shelterID, req.body)
+    await updateShelterByID(req.app.get('db'), req.params.shelterID, req.body)
       .then(() => {
         res.sendStatus(200);
       })
@@ -81,7 +81,7 @@ const updateShelter = async (req, res, next) => {
 const deleteShelter = async (req, res, next) => {
   log.debug("Calling deleteShelter...");
   //ShelterID is ON DELETE CASCADE, If a shelter is removed, those PET records associated with the ShelterID will be removed automatically
-  await deleteShelterByID(req.params.shelterID)
+  await deleteShelterByID(req.app.get('db'), req.params.shelterID)
     .then((dbResponse) => {
       if (dbResponse.affectedRows == 0) res.sendStatus(404);
       else if (dbResponse.affectedRows == 1) res.sendStatus(204);
@@ -101,7 +101,7 @@ const getPets = async (req, res, next) => {
   log.debug("Calling getPets...");
 
   let foundShelter = false;
-  await getShelterByID(req.params.shelterID)
+  await getShelterByID(req.app.get('db'), req.params.shelterID)
     .then((dbResponse) => {
       if (dbResponse.length == 0) res.status(404).send("Shelter not found");
       else foundShelter = true;
@@ -114,7 +114,7 @@ const getPets = async (req, res, next) => {
 
   if (foundShelter) {
     log.debug("Looking for pets for shelterID == " + req.params.shelterID);
-    await getAllPets(req.params.shelterID)
+    await getAllPets(req.app.get('db'), req.params.shelterID)
       .then((dbResponse) => {
         res.status(200).send(dbResponse);
       })
